scripts/generate_tasks.py
"""Utility to generate task-family boilerplate code from YAML definitions."""

from pathlib import Path
from typing import Any

import yaml


def main():
    config_path = Path(__file__).parent.parent / "config" / "task_families.yaml"
    if not config_path.exists():
        print("no task_families.yaml found")
        return
    data = yaml.safe_load(config_path)
    # placeholder; a real script might emit Python classes or test cases
    print("Loaded", len(data.get("task_families", [])), "task families")


if __name__ == "__main__":
    main()
