requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=0.21.0
requests>=2.31.0

# LLM & AI
anthropic>=0.7.0

# Testing & Evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Web framework (for dashboard)
flask>=3.0.0
flask-cors>=4.0.0

# Data & Configuration
pyyaml>=6.0
pandas>=2.0.0

# Utilities
click>=8.1.0
colorama>=0.4.6
tqdm>=4.66.0
loguru>=0.7.0
prometheus_client>=0.24.0

# Type checking
mypy>=1.4.0
types-requests>=2.31.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
